+++ changelog.rst
@@ -1,12 +1,19 @@
 Changelog
 =========
 
-Last change: 22-FEB-2024 MTS
+Last change: 06-JUN-2024 MTS
 
-0.6.9
------
+0.6.9 - 0.6.10
+--------------
 - Added the option to draw polygon picks in Picasso: Render
 - Save pick properties in Picasso: Render saves areas of picked regions in nm^2
+- Calibration .yaml file saves number of frames and step size in nm
+- ``picasso.lib.merge_locs`` function can merge localizations from multiple files
+- Mask dialog in Picasso: Render saves .png mask files
+- Mask dialog in Picasso: Render allows to save .png with the blurred image
+- Picasso: Localize - added the option to save the current view as a .png file
+- Picasso: Render - functions related to picking moved to ``picasso.lib`` and ``picasso.postprocess``
+- Picasso: Render - saving picked localizations saves the area(s) of the picked region(s) in the metadata file (.yaml)
 
 0.6.6 - 0.6.8
 -------------